+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Refresh playgrounds when filters change
+
 ## [0.18.0] - 2025-03-14
 
 ### Added